Centralize free functions in concurrency.rs using CurrentBackend type alias
