+++ ractor/src/concurrency/tokio_primitives.rs
@@ -72,74 +72,6 @@ impl super::ConcurrencyBackend for TokioBackend {
     }
 }
 
-/// Represents a task JoinHandle
-pub type JoinHandle<T> = <TokioBackend as super::ConcurrencyBackend>::JoinHandle<T>;
-
-/// A duration of time
-pub type Duration = <TokioBackend as super::ConcurrencyBackend>::Duration;
-
-/// An instant measured on system time
-pub type Instant = <TokioBackend as super::ConcurrencyBackend>::Instant;
-
-/// Sleep the task for a duration of time
-pub async fn sleep(dur: Duration) {
-    <TokioBackend as super::ConcurrencyBackend>::sleep(dur).await;
-}
-
-/// An asynchronous interval calculation which waits until
-/// a checkpoint time to tick
-pub type Interval = <TokioBackend as super::ConcurrencyBackend>::Interval;
-
-/// Build a new interval at the given duration starting at now
-///
-/// Ticks 1 time immediately
-pub fn interval(dur: Duration) -> Interval {
-    <TokioBackend as super::ConcurrencyBackend>::interval(dur)
-}
-
-/// A set of futures to join on, in an unordered fashion
-/// (first-completed, first-served)
-pub type JoinSet<T> = <TokioBackend as super::ConcurrencyBackend>::JoinSet<T>;
-
-/// Spawn a task on the executor runtime
-pub fn spawn<F>(future: F) -> JoinHandle<F::Output>
-where
-    F: Future + Send + 'static,
-    F::Output: Send + 'static,
-{
-    <TokioBackend as super::ConcurrencyBackend>::spawn(future)
-}
-
-/// Spawn a task on the executor runtime which will not be moved to other threads
-pub fn spawn_local<F>(future: F) -> JoinHandle<F::Output>
-where
-    F: Future + 'static,
-{
-    <TokioBackend as super::ConcurrencyBackend>::spawn_local(future)
-}
-
-/// Spawn a (possibly) named task on the executor runtime
-pub fn spawn_named<F>(name: Option<&str>, future: F) -> JoinHandle<F::Output>
-where
-    F: Future + Send + 'static,
-    F::Output: Send + 'static,
-{
-    <TokioBackend as super::ConcurrencyBackend>::spawn_named(name, future)
-}
-
-/// Execute the future up to a timeout
-///
-/// * `dur`: The duration of time to allow the future to execute for
-/// * `future`: The future to execute
-///
-/// Returns [Ok(_)] if the future succeeded before the timeout, [Err(super::Timeout)] otherwise
-pub async fn timeout<F, T>(dur: super::Duration, future: F) -> Result<T, super::Timeout>
-where
-    F: Future<Output = T>,
-{
-    <TokioBackend as super::ConcurrencyBackend>::timeout(dur, future).await
-}
-
 macro_rules! select {
         ($($tokens:tt)*) => {{
             tokio::select! {
